Harden checks for munge
Separate the munge checks in two commands, one to generate the key and
another to decode it, instead of using a shell pipe in a subprocess.

Also, use the default user for the check. This makes the system check if
munge is working for root instead of slurm_user. It should not be a
problem and it fixes the check when we can't sudo to another user.