Use Omnivector PPA for installing Slurm
This patch adds support for Omnivector Ubuntu PPA to install Slurm from.

By default, it uses the stable ppa:omnivector/osd and it possible to use
a custom PPA (e.g. ppa:omnivector/osd-testing) or event a URL for a repo
(as long as there's no need to verify pgp keys).