Allow for unix socket connections
In a deployment with a highly available MySQL databsae, the slurmdbd
service will use a socket connection to connect to the mysql-router.
This requires that the StorageHost and StoragePort are not set in the
slurmdbd.conf file as the MySQL connection uses the MYSQL_UNIX_PORT
environment variable in order to use a unix socket.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>